Finished end game story generator and break it off into its own section
